refactor: BaseEntity 추가 및 공통 필드(id, createAt, updateAt) 추출
